Fixing up structure of index and bin
