+++ package.json
@@ -2,7 +2,7 @@
   "name": "clevis",
   "version": "0.1.0",
   "description": "",
-  "main": "bin.js",
+  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
